remove compilation error

+++ app/sorting/Main.hs
@@ -5,5 +5,5 @@ import Control.Monad
 
 main :: IO ()
 main = do
-  [_, ls] <- forM [1,2] $ const getLine
+  [_, ls] <- forM [1,2::Int] $ const getLine
   putStrLn . sorting . map read $ words ls
